+++ release_notes.md
@@ -1,3 +1,49 @@
+Template
+=============================================================
+x.x.x Release notes (yyyy—MM-dd)
+
+Objective-C
+-----------
+The Objective-C API has been updated and your code will break!
+
+### Bugfixes:
+
+* None.
+
+### API breaking changes:
+
+* `???`
+
+### Enhancements:
+
+* `???`
+
+=============================================================
+0.5.0 Release notes (yyyy—MM-dd)
+Objective-C
+-----------
+The Objective-C API has been updated and your code will break!
+
+### Bugfixes:
+
+* None.
+
+### API breaking changes:
+
+* `getTableWithName:` renamed to `tableWithName:` in `TDBTransaction`.
+* `addColumnWithName:andType:` renamed to `addColumnWithName:type:` in `TDBTable`.
+* `columnTypeOfColumn:` renamed to `columnTypeOfColumnWithIndex` in `TDBTable`.
+* `columnNameOfColumn:` renamed to `nameOfColumnWithIndex:` in `TDBTable`.
+* `addColumnWithName:andType:` renamed to `addColumnWithName:type:` in `TDBDescriptor`.
+* unified `get` and `set` in `TDBRow`
+
+### Enhancements:
+* 
+
+
+=============================================================
+0.4.0 Release notes (2014-03-26)
+
 2014-03-26
 ==========
 + Experimental support for pinning transactions on Context.
@@ -5,7 +51,7 @@
 
 2014-03-25
 ==========
-< BREAKING CHANGE: TDBMixed removed. Use id and NSObject instead.
+! BREAKING CHANGE: TDBMixed removed. Use id and NSObject instead.
 ! BREAKING CHANGE: insertEmptyRow has been removed from table. Use insertRow:nil atIndex:index instead.
 
 2014-03-24